refactor: Tracker_ArtifactFactory dependency in ArtifactCreator
Replace `Tracker_ArtifactFactory` dependency by `TrackerArtifactCreator`
in `\Tuleap\Tracker\REST\Artifact\ArtifactCreator`.

No functional changes (artifact modal, creation of test campaign, …).

This decreases the number of usage of
`\Tracker_ArtifactFactory::createArtifact` so that we can remove it in
future contribution.

Part of request #36805: Kanban item not found when creating a card in Kanban

Change-Id: I2c1cc0c67b29cac76214b0b31287ce0f30193fe8