fix: Optimize SQL query when filtering PRs with labels
The existing query works but is heavy as it requires to compare all the
PRs of the instances against a set of labels. The changes made in this
contribution allows the DB engine to use the indexes of the PRs and
labels tables.

No functional changes expected.

Part of story #35825 Pull request new home page - Filter on labels

Change-Id: I5a98e58d48fcf8eccc90af09e8eaf5470411ba67