fix: Correct assertion parameter order and fix flakyness for TaskTrackerTest
No functional change.

$ make tests-db SEED=1707503619

The command above should pass

Why?
- in PHPUnit assertions, the expectation always comes first in the
parameters. I took the opportunity to also add return types, to use
stricter assertions and to use self instead of $this.
- the flakyness stems from the test expecting a certain order in results
from the DB query. For example, it expected to find a reminder on "due
date" at index 5 and 6, but sometimes it was switched with reminders on
"end date" (expected at index 3 and 4). The DB query to retrieve all the
reminders has an order by id, but I guess on certain conditions the
insertion order can change.

Since we do not care about the order from a business rules perspective,
I changed the test so that it filters the reminders to find the one it
wants and run assertions on it. The order of reminders can change, but
the test will keep working.

fixes request #36821 TrackersV3ToV5\TaskTrackerTest is flaky

Change-Id: I93ee3bab55ae0e454dabc63fdd70db46105d3956