refactor: Replace fields_data by InitialChangesetValuesContainer
This is a preparatory step in order to be able to move transaction from
ArtifactCreator to TrackerArtifactCreator. To achieve that we need to
move `ArtifactCreator::addReverseLinks` which depends on
InitialChangesetValuesContainer, so we need to inject this dependency
first in parameters.

Only the various calls to `TrackerArtifactCreator::create` are updated.
Those calls were not using artifact links, so `Option::nothing` is given
for `InitialChangesetValuesContainer::artifact_links`.

No functional changes.

Part of request #36805: Kanban item not found when creating a card in Kanban

Change-Id: I0de0e969e52e4c11dd212b7a1a4b25f3c24b0c94