refactor: InitialChangesetValuesContainer in createFirstChangesetNoValidation
That way the transaction will have access to the object instead of a
random associative array. This will allow us to manage reverse links in
this transaction later (See `ArtifactCreator::addReverseLinks`).

No functional changes.

Note: testItThrowAnExceptionWhenFieldDoesntExist is now removed because
not anymore relevant since long time ago: it doesn't throw, and the
assertion on warning log has nothing to do with field validation. My
guess is that the validation of the fields have been moved elsewhere and
this is a leftover (didn't dig deep enough to assert that however).

Part of request #36805: Kanban item not found when creating a card in Kanban

Change-Id: I6d710e9c1e27afe1dd0469ed31adaa3f8602bb32