Feature flag + Add the route which handle the file project upload.
Part of story #36484 create a new project from an XML template

To ease the review, the upload which use Tus will be done in a dedicated contribution.
We need to save the file to upload information to be able uploading it after.

How to test:
 - Enable the feature flag => `tuleap config-set feature_flag_create_from_custom_archive 1`
 - make dev-forgeupgrade && make composer && make dev-clear-cache
 - Play with POST project_files REST route in API Explorer (no need to
  upload file)
=> The route return the upload href
=> The 'file' stored in DB should have the right name and size given,
the user_id should be the id of the user who make the request.

Change-Id: I4b4504944f6fccb46c25a21b3bb6518285f38ae5