refactor: addReverseLinks in dedicated object
In order to be able to move addReverseLinks from ArtifactCreator to
TrackerArtifactCreator, we need first to extract it in a dedicated
object.

This decreases responsibilities of ArtifactCreator, and allow us to have
a better test coverage of this new object.

No functional changes.

Part of request #36805: Kanban item not found when creating a card in Kanban

Change-Id: I2d6e9cb35ee6bc4c8b10e3ab625c2315f7666d79