refactor: Put the `From another project I'm admin of ` in a dedicated componenent
This is the frist step before the creation of the `From Project Template
Upload` card in the `For Advanced Users` section.
The display of the cards should be done in their dedicated componenent since it will have several cards in `For Advanced Users` section now

No functionnal change expected.

Part of story #36484 create a new project from an XML template

Change-Id: Id569c0fbc2cb3e8441ee16fc4bd258212b758060