fix(tracker): Fix request #36834 Tracker_WorkflowUser must be able to see all artifacts
Given an update triggered by the Tracker triggers on a program increment, Tracker_WorkflowUser
must be able to perform this.

More globally, Tracker_WorkflowUser must be able to see all artifacts on the platform given
it's able to update them.

According to Marie-Ange, here are the steps to reproduce the issue:

You should have cross tracker triggers. The easiest way is to use Program Management plugin
There must be a hierarchy Program Increment > Iteration

In the PI tracker have a cross-tracker trigger. It should be set the following way:

  When at least one child of type Iteration has its field Status set to Planned
  Then the tracker field Status will be set to Planning

Create a PI with an open status value different than planning. For exemple, create a PI with a developping status
Open the PI, then create an iteration
=> an error will be raised, because the workflow user won't have the permission to update the status of the PI

Change-Id: I7726af228e951d8b62879206a8dbff08a61cb913