refactor: Group Float field invalid checks together
part of story #10710 search on fields with duck typing

No functional change. CI should be happy.

Why?
This is the beginning of the refactoring to transition away from invalid
checks being grouped by comparison type (equals, lesser than, etc.).
We want to group it by field type instead, as it is easier to
understand. Most of the time, we work "field by field", so it makes more
sense to have all the business rules for a given field in a few files,
instead of having all the business rules for "equals" comparison being
grouped together.

Change-Id: I1e0dc2706da16a3bd43bb4b0073470698355a0bd