feat: Prevent usage of outgoing client requests in transaction
When we are in the middle of a transaction, we don't want a request to
be performed to an external service. Transaction can be rollbacked or
transaction will take time to commit, external service should based its
behavior on committed data.

No functional change.

If you try to add a transaction in `ArtifactCreator::create`, then the
creation of the artifact via the modal in kanban should fail.

This remove the last enclosing transactions. It also moves
addReverseLinks method call outside of the transaction since it start
its own transaction.

Part of request #36805: Kanban item not found when creating a card in Kanban

Change-Id: Ib8fd0db2fd417067d512cc0ef9ab66705404b947