chore: Unit test invalid comparisons and values for float
part of story #10710 search on fields with duck typing

No functional change. Unit tests should pass.

Why?
This will help us refactor the architecture of the invalid checks. We
want to transition away from the current state where validation is
grouped by comparison type (equals, lesser than, etc.). We want to group
it by field type instead, as it is easier to understand. Most of the
time, we work "field by field", so it makes more sense to have all the
business rules for a given field in a few files, instead of having all
the business rules for "equals" comparison being grouped together.

Change-Id: I1c051814b7b8f07d70426eb74e4402cc021b408b