fix: Ease access to Submit and stay button
"Submit and stay" or "Submit and continue" are currently hidden in a
dropdown (dropup) menu. This can be frustrating when you use this
feature a lot to have to click twice to submit the artifact (one to open
the menu, one to submit and stay|continue).

This contribution displays all buttons by default, and fallback on
dropdown menu on small screens.

Closes request #36839: Submit and Stay option with one click

Change-Id: I032c3ee1f6644a73fd79cfc7066769687646ccb0