Close request #36840 Cannot access to tuleap container from "dev" e2e test
No functional change. If you run `make tests-e2e` and edit your
`/etc/hosts`, you can access the `https://tuleap/` web page. The cypress
app can also access it.

Why?
It might be due to a change in my version of Docker. Since the "tuleap"
network is internal, maybe it really is enforced now and I cannot reach
it from the host. Switching the tuleap network to "not internal" solved
this.

Change-Id: I16f5431d9c39ca53779cf9ab1f3027a9164e3abb