refactor: Use Flat field checker for cross-tracker too
part of story #10710 search on fields with duck typing

No functional change. CI should be happy.

Why?
It allows us to drop the previous "visitFloat" methods. Unit tests
should check that nothing is broken. FloatFieldChecker is merged with
FlatFloatFieldChecker so that all the validation logic is in the same
file for float fields.

Change-Id: Iee4be66273f20f1356474ef4950201285bdd6d35