fix: Close HTTP connections when running the REST test suite
Currently the connections stay open and since we are running the whole
test suite in a single process it means the file descriptors are kept
for the whole test suite.

Depending on the default settings of your OS/user/Docker container it is
possible the reach the limit on the number of file descriptors opened.
This is currently the case on NixOS 23.11 but it is likely the issue is
visible elsewhere (or will be at some point anyway).

No changes for situation where the test suite was already running fine.

Part of request #35459: Bump CI agents to NixOS 23.11

Change-Id: I1393786c940b3399fb0ffdf95a2f9163a3c8e9b6