Fix: remove NIH syndrom in our dates computation in Roadmap
In some configurations, for example when the Tuleap server/user timezone
is Europe/Paris but the user's operating system is Atlantic/Azores then
the months displayed in the Roadmap widget are wrong.

Timezones are hard, we should not try to be smarter than dedicated
libraries. Replace the default javascript Date usage by [luxon][0]. We
now force the timezone to be the one set in Tuleap user preference page,
so that we have consistent results among dates computed on client and
server side.

To test: change your OS timezone, and refresh the Roadmap widget. The
today indicator should always be on February.

[0]: https://moment.github.io/luxon/

Closes request #36845: The Current Date Marker on the Road Map is off by 1 Month

Change-Id: Iadc25ebdd8cb91cc61be989dd2b7d87f122cfc6e