fix: display the card "From user project" description when the card is not selected
Part of

This bug was not annoying before there is only one card in the `Advanced` option.
When the new card "From project file" will be introduced this visual bug would be more annoying.
This behaviour is also more consistent with the behaviour of the tracker
creation view.

How to test:
 - Create a new project and select the `From another project I'm admin
   of` card
=> The list of the project is displayed in a list picker
 - Select another card and check the `From another project I'm admin of`
=> The card is unselected and the description is displayed again.

=> The creation of the project should still works as before.

Change-Id: Ia8367228701dd678c8bee5c5b78f03b639c5a52e