[chore] prepare merge of POST project_files route inside POST projects
part of story #36484 create a new project from an XML template

Why merging route?
Previous architecture was weird,
uploading a file is conceptually the same thing than creating a project from an xml template

the route POST should handle the creation of the project, we need to merge the two routes

Why this refacto?
In order to merge the project_files into the existing POST project route,
I'll need to add a new ProjectCreationData from zipArchive

This patch:
 - adds types in ProjectCreationData
 - add namespace in ProjectCreationData
 - moves file concerning Project creation inside a common namespace src/common/Project/REST/v1/Project/

How to test the absence of #regression:
 - you can create a project from any template from UI
 - you can create a project from a project you're admin of in advanced settings
 - in the Details tab of the created project, the template used by project is correct

Change-Id: Ia963db0f37fc90216bab9fc1693db45d6943b3c3