fix: Adjust all icons of cross-tracker widget
fixes request #36856 Save icon no longer shown in cross-tracker search
widget

How to test:
When you modify a cross-tracker report, for example by editing the query
or choosing another tracker, then the "save the report" button should
have a "save" icon.
Other icons should still show up.

Why?
aria-hidden attribute is added to all icons except the "cross" icon to
remove a tracker from the list. This one should not be an icon but
should be a button, but that is outside the scope of this patch. This
attribute hides the icon from screen readers, because they are here for
"cosmetic" purpose only.
Set "fa-solid" class for all icons missing it.
Also added a margin to the "info" icon that was glued to the text.

Change-Id: I162465de34100cc190f7ac21e42f1ba95d134147