chore: Migrate rest-querier and store to TypeScript
part of request #35894 Convert timetracking-overview-widget to Vue3 + TS

Note:

The current strategy is to convert the store and the rest querier to TS
first in order to identify things to improve in the future. It allows us
to have a first typecheck command on TS files only.

The next steps are:
- Move to vue 2.7.x and migrate components to composition API step by
  step.
- Once done, bump to vue3.
- Move components and their tests to TS + use vue-tsc to typecheck the app.

[Bonus steps]
- Move from tlp-fetch to fetch-result
- Use Option when there are nullable properties
- Move from webpack/jest to vite/vitest

How to test:
--> No functional change expected
--> CI is happy

Change-Id: Icd10a146fdcadfaec43b3c43d900270df3b8e3f2