refactor: remove useless transaction for PATCH kanban/:id/items
There is already a transaction in
`\Tuleap\Tracker\Artifact\Changeset\NewChangesetCreator::create`, we
don't need an outter transaction when we move a bunch of artifacts in a
column.

This applies also to PATCH kanban/:id/backlog and PATCH
kanban/:id/archive.

No functional changes.

Part of request #36805: Kanban item not found when creating a card in Kanban

Change-Id: I72da4a3c34a983536368c71bdc742e1da916ccae