fix request #36867 Test execution is not updated when using step buttons
When updating a test with steps, the test is not updated in local and
its `previous_result` attribute still the same. So when we compute the
score, the user granted to pass this test is not the good user but the
one who updated the test previously (the one who created the campaign
for example)

*Testing:*

In a test campaign created by userA, update a test with step with userB.
When the whole test goes passed, the updated score should be added to
userB. You can test with 2 tab (one per user) to check the score update.

Change-Id: I46854fd9b0db353f3d9b5f113c8a983c05dd9d4f