refactor: extract minimal tlp-styles as a lib
part of request #29982 Split `tlp` in smaller packages

No functional change. TLP styles should be the same as before.

Why ?
- Parts of TLP that were imported from elsewhere are now moved to a
dedicated tlp-styles lib. It is a first step towards moving as much of
the styles as possible to the lib.
- @import is deprecated, replace it by @use everywhere possible. The
last remaining ones are for FontAwesome. Their documentation still
advises to include it with @import. We can deal with it separately.
- colors are defaulted to orange and passed by each theme with `@use
with()`
- Prefix placeholders and mixins with `-` to mark them as private. Those
placeholders are never meant to be reused outside of their stylesheet.

Change-Id: I6cf7c30b3541b1e19a94fcea31d7896fcbe20e5f