refactor: filtering and templating callback for all kinds of users
part of story #35822 Pull request new home page - Filter on Reviewer

In this contribution, the filtering and templating callbacks of the
Author selector entry are extracted from the Author filter context.

It will allow us to reuse them for the Reviewer filter, since it is also
a user selector, the display and filtering being strictly the same.

-- no functional change expected

Change-Id: I088f548e166be3e74392e59ae234fe5ef98594c5