feat: Send 404 if project is not active nor deleted nor suspended
If a project is not active, not deleted and not suspended, we should
send a 404 since its status does not lead currently to a well formed
project.

This prepares the field for XML import of project, where we plan to have
an intermediate status until the import is complete (ex: status=X). You
can try this feature by updating manually the database and setting the
status of a project to 'X' (in `groups` table). `GET /projects/:id`
should return a 404 and not a 403.

Part of story #36484 create a new project from an XML template

Change-Id: I94b7c8e374d9141465d9640ab03163b6023f9e94