feat: Create project before sending tus url
To not loose project creation context when we send back the url to
upload XML archive via tus protocole, we need to create the project
first.

It is created in a new status 'X' (as in XML, if you have a better
suggestion…).
* This project is not accessible
* This project does not appear in pending projects page

In next contribution the id of this new project will be stored alongside
tus file in `project_file_upload` table so that we will be able to
continue the XML import in the Finisher.

Note: In order to be consistent with other resources (e.g docman files),
we send back the id and uri of the project alongside the upload href.
This is needed for the client to know where to look at when the project
will be fully created (for now it will get only 404 since XML import is
not implemented yet).

part of story #36484 create a new project from an XML template

Change-Id: I01a3cacd1c8408653ca8cb5dc290a2eaef45e443