fix: Change selection when a date is chosen manually
Part of request #11352 Timetracking should be able to compute time for one day

When users select values manually, the preset selectbox still display the last selected value and it is rather misleading because it is not clear wheteher it is the manually selected dates or the preset that has been applied.
Displaying a placeholder in this case seems to be the best solution.
This commit changes this behavior and now a new 'Please choose' is written in the selection option when a date is selected in the date picker.

Change-Id: I838233975cbd297a70f78e74a0cda7315db54734