feat: store project id alongside tus information
The id of the project will be needed in order to complete the setup once
the upload of the archive is finished.

You can test by using the api explorer and `POST /projects` from
archive. This will return a project id and an upload href. This id is
present in the corresponding row in `project_file_upload` table.

part of story #36484 create a new project from an XML template

Change-Id: If3ae8e8df53dd1994ab1c6efb8c8bba695bb1a82