chore: Add "permissions" DB tests on Numeric
part of story #10710 search on fields with duck typing

No functional change.

Why?
- Remove "Invalid" tests, such as "user cannot ready any field",
"invalid comparison", "comparison to empty", "comparison to myself".
They are already covered at unit-test level, which is much faster than
DB tests.
- Replace the epic tracker and outsider user by "Permissions" test cases
similar to the rest of the DB tests. It's easier to understand when all
the test files work the same way.

Change-Id: If010e373eded15944d55cb909bd95fdfeccf736f