feat: upload the archive file and "create" the project.
At this step, the file is uploaded and the created project has the status X
For now, you will be redirected to the page `This project is not active` when the project is created form an archive.The dedicated page will be done after.
Project creation from others source should still work has before.

To test:
 - Enable the feature flag: `tuleap config-set feature_flag_create_from_custom_archive 1`
 - Create a new project, select the `From a project template` card,
   select an archive, then click on the next button.
 - Open the web console.
 - Fill the form and create the project
=> you are redirected to the project but `This project is not active`` will
be displayed
=> In the DB the project is saved with the status `X` in the `groups` table, the file info is saved
=> The information about the file upload can be see in the web console
in `project_file_upload` table.
- go into the Tuleap web container => `make bash-web`
- Go where the file is uploaded, ( cd /var/tmp/project/ongoing-upload/)
=> You can see the file id name from `project_file_upload`

Part of story #36484 create a new project from an XML template

Change-Id: I5e7b6bfc7cb4ec32544eb0a95af6bb2f10c8ebf0