feat: import XML archive into new project
Initiate the creation of a project via REST api (from archive). With the
given upload_href, upload (with tus client) the given archive.

Then as site admin go to the following url, replacing <id> with the
project id you've just created:
https://tuleap-web.tuleap-aio-dev.docker/project/:id/admin/services

The enabled services should match the one in the archive, and content is
imported as well (trackers, artifacts, …).

Note: the project is only accessible to site admin for now, we will deal
with the its activation in a dedicated contribution.

Changes in `ProjectXMLImporter.class.php` are needed to make sure that
project does not have enabled services from default template (in
addition to the ones of the archive).

part of story #36484 create a new project from an XML template

Change-Id: I71ce61541cb34f61baddad1dc6d638de9ace1f74