fix(bug): Close request #37207 Cannot retrieve Document file versions if there are more than 100
Given a file document that have more than 100 versions, then the UI will not display any of them.

This is due to REST error when the offset is beyond 50.

It seems that argument order matters for RESTler, so that the previous
{@max 50} was applied to $offset and not $limit ...

For backward compatibility, the max limit has been raised to 100.

How to test
----------

* Query the REST route GET docman_files/:id with an offset > 50

=> The route must return the result as expected.

Change-Id: I54fdfcbb6d013d66020f2ade254bee5bbb175f46