feat: import of project archive is now async
The extraction and import of the archive is now delegated to a worker
event. This is necessary in order to later not be blocked by a timeout
or a loss of db connection if the archive is big.

To test, Create a project from XML, the project should be created after
the worker events are processed.

A check on worker availability will be added in dedicated contributions.

part of story #36484 create a new project from an XML template

Change-Id: Idc7205a6fd4e0a448ec5904cdaa1a1e5c8e62da2