feat: XTS support open lists for all binds
Part of story #10710 search on fields with duck typing

Last part of XTS duck typing: open lists. User is now able to query on
list fields with any bind even if it's an open list.

*Testing:*

With some trackers with open list fields, performs a cross tracker
search on these trackers with any operator on open list fields. It
should works:
- if there is no open list
- if all fields are open list
- if some fields are open list

Change-Id: I4d3d2f169014224628a65dee92f8aab999354f88