feat: Check worker availability for xml import
Since creation of a project with an archive is now delegated to an
asynchronous worker, we need to be sure that workers are available
before allowing users to create project with this option.

To be able to test you have to set 0 for sys_nb_backend_workers in
forgeconfig table (there is a high chance that you cannot do it via
tuleap config-set due to protections that have been put in place).
Once set to 0, creation from project cannot succeed, and is not proposed
as an option.

part of story #36484 create a new project from an XML template

Change-Id: I64c07c64726b6093b799328b0acb747273740297