refactor: Drop InvalidFieldChecker interface
part of story #10710 search on fields with duck typing

No functional change. CI should be happy.

Why?
Now that all the validation rules are grouped by type of field, the goal
is to remove another level of indirection:
  $parameters->getCheckerProvider()->getInvalidFieldChecker()->check().
Due to the way the Field Visitor works, we cannot pass parameters to it.
This means we cannot pass the current Comparison to the check method
directly in the "accept*" methods. This is why the CollectorParameters
included a class that returns the checker, instead of just doing the
check immediately. Hence the existence of the
IProvideTheInvalidFieldCheckerForAComparison and InvalidFieldChecker
interfaces.

In order to work around this problem, we replace the Field Visitor by a
match on field classname. It breaks again the "visitor" spirit, but also
has advantages:
- default branch of match expression should handle all new fields as
"external" fields (this was not the case in the visitor which required
adding a method)
- we can pass different parameters to each branch. With the visitor,
there are no parameters (hence the additional indirection)
- it is more concise code
- InvalidSearchableCollectorParameters no longer needs to depend on
field "Checker" class. The FlatInvalidFieldChecker can be injected at a
higher level.

Change-Id: Id48bded6acd455b4fb37bd3631203bef005d1116