feat: Activate project after archive import
Once the archive is successfully imported, the project is "activated"
=> if platform does not require administrators approval, the project is
active, site admin receive an email, project admins receive an email.
=> if platform requires administrators approval, the project is pending,
site admin receive an email

A confirmation email will be sent to project admins when the archive
import is successful, so that they can now that its approval is ongoing.
This will be done in a dedicated contribution.

In order to be able to be activated, the import should properly end
first.  Currently it is stopped since async worker does not have a
current user, and therefore there is a redirect (!) deep down in the
stack trace (See `\AgileDashboard_XMLController::importProject`,
`checkUserIsAdmin`). Now the user id who uploaded the archive is now
part of the worker payload, but to retrieve it we need to pass the
request to the finisher, we end up updating every usage of
`\Tuleap\Tus\TusFinisherDataStore::finishUpload`.

Part of story #36484 create a new project from an XML template

Change-Id: Ib99e5d22da12cb170f08e472da2a2876f4c84ca4