chore: Close request #37224 Set timeout to 30s for project creation in e2e tests
No functional change.

Why ?
On heavily loaded CI environments (such as for the Nightly job), 20
seconds was very close to the actual time. This means we can sometimes
run into the timeout even though the test is working. As project
creation is known to take some time, even in dev environments, it should
not harm us to push the timeout to 30 seconds.

Change-Id: I51d432b9aa3ce85641c2e800b7ab85adeeab0536