feat: Indicate in project details if created from archive
When a project is created from an uploaded archive, we should indicate
that in project details (both project admin, and site admin parts).

It displays a link "User provided template". For now this link is dummy,
since we currently lose the archive at the end of the import. A
dedicated task to move it to a safe place, and a controller to handle
the download will be done in dedicated contributions.

Part of story #36484 create a new project from an XML template

Change-Id: Id71015ba684904a9c947046b09b8d61c99e43776