chore: Add unit tests for semantics in cross-tracker search
Part of story #36833 search on semantics and always there fields even if
not defined

No functional change. CI should be happy.

Notes:
The CheckMetadataUsage interface is pushed down one level. Ideally, we
should move interfaces down to the field / semantic DAOs, but due to the
presence of a cache, it's going to require more work.

Change-Id: I4b84bfb5cb5fcf5ecf9350c8c8595624ce35ebd3