perf: Avoid a full scan of the DB table when filtering PR "related to me"
Moving the filtering of this filter inside the query allows the DB to
apply all the "easy" filters before considering to look into the list of
changes. It is easier for the DB to come with an optimized plan for
this.

No functional changes.

Part of story #35823: Pull request new home page - Filter on Related to me

Change-Id: Ib07997311924c782031a0d923bc07cdefd84bb52