feat: Change label of suggestions and info text
Part of story #10710 search on fields with duck typing

The label of the suggestions selectbox is changed. The info text
underneath the query editor field is changed. They should both be
translated in French. The selectbox and the query editor should be
visually aligned.

Why ?
The goal is to have better UX.
Snapshot test is removed. Asserting the shape of HTML / CSS we do not
control (CodeMirror) can only break unexpectedly and it does not help us
to check things look the way we expect. If we controlled 100% of the
HTML / CSS, it might have some value, but here it is only going to
break. Therefore, I chose to remove the snapshot test.

Change-Id: I68ec00603bfcf76930e5a5621a713ae8c492ef2b