fix: Optimize query on @status
Part of story #36833 search on semantics and always there fields even if not defined

The query built uses many join even in the where, it was not efficient.

So replace it with a smaller one inspired by what we've done for list
fields. It should take less time to find same artifacts.

Also remove a useless join in list field : the join on changeset
brings nothing as data is already present in artifact and
changeset_value.

*No functional changes expected*

Change-Id: Icf30f6393cca68b83a1e9bfc2a1ba3c151fd079f