fix: Optimize query on @status
Part of story #36833 search on semantics and always there fields even if not defined

The query built uses many join even in the where, it was not efficient.

So replace it with a smaller one inspired by what we've done for list
fields. It should take less time to find same artifacts.

Also remove a useless join in list field : the join on changeset
brings nothing as data is already present in artifact and
changeset_value.

*No functional changes expected*

Note: This commit fix a previous broken one
3235405f177fe9e32a055aab3e7a4bae4aeef5b0 reverted in
25e7d37df1b8d73322e0f01801de3ac515802665

Change-Id: I59e57481d90bb1eca1e09bb9c95a4d47c5f66d30