fix: Remove check of zip archive in finisher
Checking the file content has nothing to do with the upload of a file in
the tus protocole. The current situation is not correct since the tus js
client does not handle well the 400 that is raised if a file is not a
zip archive (the error message is lost in the process, the retry is
broken, …). To fix the situation we now move the check to the async
worker.

Part of story #36484 create a new project from an XML template

Change-Id: I2a5123906b3ea7504cc27c100c0e976705e8154d